+++ adminalerts/tests/test_ui.py
@@ -4,11 +4,12 @@ from django.urls import reverse
 from django.utils import timezone
 
 from selenium.common.exceptions import NoSuchElementException
+from selenium.webdriver.common.by import By
 
 # Projectroles dependency
 from projectroles.tests.test_ui import TestUIBase
 
-from .test_models import AdminAlertMixin
+from adminalerts.tests.test_models import AdminAlertMixin
 
 
 class TestAlertUIBase(AdminAlertMixin, TestUIBase):
@@ -38,7 +39,6 @@ class TestAlertMessage(TestAlertUIBase):
         """Test visibility of alert message in home view"""
         expected = [(self.superuser, 1), (self.regular_user, 1)]
         url = reverse('home')
-
         self.assert_element_count(
             expected, url, 'sodar-alert-site-app', 'class'
         )
@@ -50,7 +50,6 @@ class TestAlertMessage(TestAlertUIBase):
 
         expected = [(self.superuser, 0), (self.regular_user, 0)]
         url = reverse('home')
-
         self.assert_element_count(
             expected, url, 'sodar-alert-site-app', 'class'
         )
@@ -62,7 +61,6 @@ class TestAlertMessage(TestAlertUIBase):
 
         expected = [(self.superuser, 0), (self.regular_user, 0)]
         url = reverse('home')
-
         self.assert_element_count(
             expected, url, 'sodar-alert-site-app', 'class'
         )
@@ -70,9 +68,8 @@ class TestAlertMessage(TestAlertUIBase):
     def test_message_login(self):
         """Test visibility of alert in login view with auth requirement"""
         self.selenium.get(self.build_selenium_url(reverse('login')))
-
         with self.assertRaises(NoSuchElementException):
-            self.selenium.find_element_by_class_name('sodar-alert-site-app')
+            self.selenium.find_element(By.CLASS_NAME, 'sodar-alert-site-app')
 
     def test_message_login_no_auth(self):
         """Test visibility of alert in login view without auth requirement"""
@@ -81,7 +78,7 @@ class TestAlertMessage(TestAlertUIBase):
 
         self.selenium.get(self.build_selenium_url(reverse('login')))
         self.assertIsNotNone(
-            self.selenium.find_element_by_class_name('sodar-alert-site-app')
+            self.selenium.find_element(By.CLASS_NAME, 'sodar-alert-site-app')
         )
 
 
@@ -92,12 +89,10 @@ class TestListView(TestAlertUIBase):
         """Test existence of items in list"""
         expected = [(self.superuser, 1)]
         url = reverse('adminalerts:list')
-
         self.assert_element_count(expected, url, 'sodar-aa-alert-item', 'id')
 
     def test_list_buttons(self):
         """Test existence of buttons in list"""
         expected = [(self.superuser, 1)]
         url = reverse('adminalerts:list')
-
         self.assert_element_count(expected, url, 'sodar-aa-alert-buttons', 'id')