+++ projectroles/__init__.py
@@ -10,3 +10,6 @@ del get_versions
 default_app_config = (
     'projectroles.apps.ProjectrolesConfig'  # pylint: disable=invalid-name
 )
+
+from . import _version
+__version__ = _version.get_versions()['version']